# SHIP/docs/openapi.json
{
  "components": {
    "schemas": {
      "Abort-migrationRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/Abort-migrationRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "reason": {
            "description": "Recorded on the run",
            "maxLength": 200,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DurationBySize": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/DurationBySize.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "median_rows_per_second": {
            "description": "Median backfill throughput of past runs; 0 with no history",
            "format": "double",
            "type": "number"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/DurationPoint"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "points",
          "median_rows_per_second"
        ],
        "type": "object"
      },
...
      "Lock-waitsResponse": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/Lock-waitsResponse.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LockWait"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Drives pgroll migrations through expand, verify and contract under live load, and reverts them when locks or errors exceed their thresholds. Reads need the viewer or operator token, everything else the operator token, as `Authorization: Bearer \u003ctoken\u003e`. Errors are RFC 9457 problem documents.",
    "title": "shipd: zero-downtime schema migrations",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/analytics/duration-by-size": {
      "get": {
        "operationId": "duration-by-size",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurationBySize"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "viewer"
            ]
          }
        ],
        "summary": "Migration duration against table size",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/analytics/lock-waits": {
      "get": {
        "operationId": "lock-waits",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lock-waitsResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "viewer"
            ]
          }
        ],
        "summary": "Lock wait time during each migration",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/compat": {
      "get": {
        "operationId": "get-compat",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Matrix"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "viewer"
            ]
          }
        ],
        "summary": "Compatibility matrix: application versions against schema versions, with live connections",
        "tags": [
          "compatibility"
        ]
      }
    },
    "/v1/compat/check": {
      "get": {
        "operationId": "check-compat",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "app_version",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._-]{1,20}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompatCheck"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "viewer"
            ]
          }
        ],
        "summary": "Deploy gate: may this application version be rolled out now?",
        "tags": [
          "compatibility"
        ]
      }
    },
    "/v1/migrations": {
      "get": {
        "operationId": "list-migrations",
        "parameters": [
          {
            "description": "next_cursor of the previous page",
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "description": "next_cursor of the previous page",
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageRun"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "viewer"
            ]
          }
        ],
        "summary": "List migration runs, newest first",
        "tags": [
          "migrations"
        ]
      },
      "post": {
        "description": "Returns 202 for a new run. Safe to retry: submitting the same name and operations again returns the existing run with 200. A different migration under the same name, or any submission while another migration is in flight, is a 409.",
        "operationId": "submit-migration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Submit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Accepted"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "operator"
            ]
          }
        ],
        "summary": "Submit a migration and start expanding it",
        "tags": [
          "migrations"
        ]
      }
    },
    "/v1/migrations/{id}": {
      "get": {
        "operationId": "get-migration",
        "parameters": [
          {
            "description": "Run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Run id",
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "viewer"
            ]
          }
        ],
        "summary": "Get a run with its progress",
        "tags": [
          "migrations"
        ]
      }
    },
    "/v1/migrations/{id}/abort": {
      "post": {
        "description": "Refused with 409 if a connected application version can only run on the new schema version. Returns 202 while reverting; poll the run. Safe to retry.",
        "operationId": "abort-migration",
        "parameters": [
          {
            "description": "Run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Run id",
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Abort-migrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Accepted"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "operator"
            ]
          }
        ],
        "summary": "Abort and revert a migration that has not been contracted",
        "tags": [
          "migrations"
        ]
      }
    },
...
    "/v1/plans": {
      "post": {
        "description": "Compares the desired schema with the live one and returns the expand-contract plan: pgroll operations, the steps of each phase with the locks they take, and a backfill estimate. Changes nothing.",
        "operationId": "plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Desired"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": [
              "operator"
            ]
          }
        ],
        "summary": "Plan a migration from a schema diff",
        "tags": [
          "migrations"
        ]
      }
    }
  }
}